Fix categories, assertions for control after changing response format
