+++ src/ggrc-client/js/components/tree/tree-view.js
@@ -3,7 +3,7 @@
  Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
  */
 
-import '../sort/sort-by';
+import '../sort/sort';
 import template from './templates/tree-view.stache';
 
 let viewModel = can.Map.extend({
