+++ src/ggrc-client/js/templates/controls/info.stache
@@ -140,48 +140,28 @@
                         Assertions
                         <i class="fa fa-asterisk"></i>
                       </h6>
-                      {{#if assertions}}
-                        {{#if_helpers '\n #if' instance.snapshot '\n or #if' instance.isRevision}}
-                            <ul>
-                              <sort-by is-snapshot="true" {items}="assertions">
-                                {{#sortedItems}}
-                                  <li>{{name}}</li>
-                                {{/sortedItems}}
-                              </sort-by>
-                            </ul>
-                          {{else}}
-                            <ul>
-                              <sort-by {items}="assertions">
-                                {{#sortedItems}}
-                                  <li>{{name}}</li>
-                                {{/sortedItems}}
-                              </sort-by>
-                            </ul>
-                          {{/if_helpers}}
+                      {{#if assertions.length}}
+                        <ul>
+                          <sort {items}="assertions">
+                            {{#sortedItems}}
+                              <li>{{.}}</li>
+                            {{/sortedItems}}
+                          </sort>
+                        </ul>
                       {{else}}
                         <span class="empty-message">None</span>
                       {{/if}}
                     </div>
                     <div class="span4">
                       <h6>Categories</h6>
-                      {{#if categories}}
-                        {{#if_helpers '\n #if' instance.snapshot '\n or #if' instance.isRevision}}
-                            <ul>
-                              <sort-by is-snapshot="true" {items}="categories">
-                                {{#sortedItems}}
-                                  <li>{{name}}</li>
-                                {{/sortedItems}}
-                              </sort-by>
-                            </ul>
-                          {{else}}
-                            <ul>
-                              <sort-by {items}="categories">
-                                {{#sortedItems}}
-                                  <li>{{name}}</li>
-                                {{/sortedItems}}
-                              </sort-by>
-                            </ul>
-                          {{/if_helpers}}
+                      {{#if categories.length}}
+                        <ul>
+                          <sort {items}="categories">
+                            {{#sortedItems}}
+                              <li>{{.}}</li>
+                            {{/sortedItems}}
+                          </sort>
+                        </ul>
                       {{else}}
                         <span class="empty-message">None</span>
                       {{/if}}
